# Add allocator param to all public functions to improve performance